Add tests for TrainingController.

+++ app/Http/Controllers/TrainingController.php
@@ -130,6 +130,7 @@ class TrainingController extends Controller
      */
     public function destroy($trainingId)
     {
+        $this->authorize('edit');        
         Training::findOrFail($trainingId)->delete();
     }
